requirements_freeze.txt
aenum==3.1.15
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
arxiv==2.1.0
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
curl_cffi==0.6.1
dataclasses-json==0.6.3
deprecation==2.1.0
distro==1.9.0
docstring-inheritance==2.1.2
duckduckgo_search==4.5.0
e2b==0.14.9
fastapi==0.110.0
feedparser==6.0.10
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonrpcclient==4.0.3
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.28
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langserve==0.0.46
langsmith==0.1.12
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.3
openai==1.13.3
orjson==3.9.15
packaging==23.2
pinecone-client==3.1.0
pycparser==2.21
pydantic==1.10.14
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sse-starlette==1.8.2
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0
websockets==12.0
xata==1.3.0
yarl==1.9.4
